move list_entry to core/widgets; move footer to widgets/sections
