+++ lib/core/widgets/list_entry.dart
