+++ lib/features/settings/presentation/pages/settings_page.dart
@@ -1,7 +1,7 @@
 import 'package:coffeecard/base/strings.dart';
-import 'package:coffeecard/features/settings/presentation/widgets/footer.dart';
 import 'package:coffeecard/features/settings/presentation/widgets/sections/about_section.dart';
 import 'package:coffeecard/features/settings/presentation/widgets/sections/account_section.dart';
+import 'package:coffeecard/features/settings/presentation/widgets/sections/footer.dart';
 import 'package:coffeecard/features/settings/presentation/widgets/sections/profile_section.dart';
 import 'package:coffeecard/widgets/components/scaffold.dart';
 import 'package:flutter/material.dart';
