+++ lib/features/settings/presentation/widgets/sections/footer.dart
